fix(ncf): make score float not int

+++ src/main/java/ms/gamerecommender/business/service/ai/ncf/NcfDataPoint.java
@@ -1,3 +1,3 @@
 package ms.gamerecommender.business.service.ai.ncf;
 
-public record NcfDataPoint(int userId, int gameId, int score) { }
+public record NcfDataPoint(int userId, int gameId, float score) { }
